+++ Editor/Source/EditorLayer.h
@@ -2,6 +2,7 @@
 
 #include "Core.h"
 #include "Panel/PanelSystem.h"
+#include "EditorCamera.h"
 
 namespace Core
 {
@@ -11,6 +12,8 @@ namespace Core
         bool Fullscreen = true;
         ImGuiDockNodeFlags Flags = ImGuiDockNodeFlags_None;
         ImGuiWindowFlags WindowFlags = ImGuiWindowFlags_MenuBar | ImGuiWindowFlags_NoDocking;
+        ImVec2 ViewportLeftTop;
+        ImVec2 ViewportRightBottom;
     };
 
     struct EditorState
@@ -18,8 +21,8 @@ namespace Core
         PanelInformation PanelInfo;
         PanelSystem Panels;
         DockspaceSettings Dockspace;
-
         ImVec2 LastFrameViewport;
+        EditorCamera Camera;
     };
 
     class EditorLayer : public Layer