+++ Engine/Source/Core/Layer/ImGuiLayer.cpp
@@ -22,9 +22,7 @@ namespace Core
         io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;
 
         // TODO: Check if path is valid
-        // io.Fonts->AddFontFromFileTTF("EngineResources/Fonts/Open_Sans/static/OpenSans-Bold.ttf", 12.0f);
 
-        // Styling
         ImGui::StyleColorsDark();
 
         // Will be started by the application, after the Engine::SetupWithApplication call
@@ -33,34 +31,6 @@ namespace Core
 
         // TODO: Make CONFIGURABLE and SAVABLE
         auto &colors = ImGui::GetStyle().Colors;
-        // colors[ImGuiCol_WindowBg] = ImVec4{0.1f, 0.105f, 0.11f, 1.0f};
-
-        // // Headers
-        // colors[ImGuiCol_Header] = ImVec4{0.2f, 0.205f, 0.21f, 1.0f};
-        // colors[ImGuiCol_HeaderHovered] = ImVec4{0.3f, 0.305f, 0.31f, 1.0f};
-        // colors[ImGuiCol_HeaderActive] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-
-        // // Buttons
-        // colors[ImGuiCol_Button] = ImVec4{0.2f, 0.205f, 0.21f, 1.0f};
-        // colors[ImGuiCol_ButtonHovered] = ImVec4{0.3f, 0.305f, 0.31f, 1.0f};
-        // colors[ImGuiCol_ButtonActive] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-
-        // // Frame BG
-        // colors[ImGuiCol_FrameBg] = ImVec4{0.2f, 0.205f, 0.21f, 1.0f};
-        // colors[ImGuiCol_FrameBgHovered] = ImVec4{0.3f, 0.305f, 0.31f, 1.0f};
-        // colors[ImGuiCol_FrameBgActive] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-
-        // // Tabs
-        // colors[ImGuiCol_Tab] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-        // colors[ImGuiCol_TabHovered] = ImVec4{0.38f, 0.3805f, 0.381f, 1.0f};
-        // colors[ImGuiCol_TabActive] = ImVec4{0.28f, 0.2805f, 0.281f, 1.0f};
-        // colors[ImGuiCol_TabUnfocused] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-        // colors[ImGuiCol_TabUnfocusedActive] = ImVec4{0.2f, 0.205f, 0.21f, 1.0f};
-
-        // // Title
-        // colors[ImGuiCol_TitleBg] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-        // colors[ImGuiCol_TitleBgActive] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
-        // colors[ImGuiCol_TitleBgCollapsed] = ImVec4{0.15f, 0.1505f, 0.151f, 1.0f};
 
         colors[ImGuiCol_WindowBg] = ImVec4{0.13f, 0.14f, 0.16f, 1.0f};
 
@@ -118,5 +88,12 @@ namespace Core
             ImGui::RenderPlatformWindowsDefault();
             glfwMakeContextCurrent(backup);
         }
+    }
+
+    void ImGuiLayer::SetFont(const char *path, float size)
+    {
+        ImGuiIO &io = ImGui::GetIO();
+        (void)io;
+        io.Fonts->AddFontFromFileTTF(path, size);
     }
 }