+++ Engine/Source/Platform/Platform.h
@@ -1,10 +1,12 @@
 #pragma once
 
 #include "Base.h"
+#include <string>
 
 namespace Core
 {
-    class Platform
+    // TODO: maybe no CE_API, maybe platformAPI that does the same thing expect without having full access to platform code (sucks but i get it :))
+    class CE_API Platform
     {
     public:
         Platform() {}
@@ -30,6 +32,11 @@ namespace Core
         static void *MemCpy(void *Block, const void *Source, CeU64 size);
         static void Free(void *Memo);
         // ------------
+
+        // --------------- FILE DIALOGS ------------------
+        static std::string OpenFileDialog(const char *filter);
+        static std::string SaveFileDialog(const char *filter);
+        // ------------------------------------------------
     };
 
 }