+++ Engine/Source/Platform/PlatformWin32.cpp
@@ -2,6 +2,10 @@
 
 #if CE_PLATFORM == CE_WIN32_PLATFORM_VALUE
 #include <windows.h>
+#include "Core/Engine.h"
+#include <GLFW/glfw3.h>
+#define GLFW_EXPOSE_NATIVE_WIN32
+#include <GLFW/glfw3native.h>
 
 namespace Core
 {
@@ -76,6 +80,55 @@ namespace Core
     {
         free(Memo);
     }
+
+    std::string Platform::OpenFileDialog(const char *filter)
+    { 
+        OPENFILENAMEA ofn;
+        CHAR szFile[260] = {0};
+        CHAR currentDir[256] = {0};
+        ZeroMemory(&ofn, sizeof(OPENFILENAME));
+        ofn.lStructSize = sizeof(OPENFILENAME);
+        ofn.hwndOwner = glfwGetWin32Window((GLFWwindow *)Engine::GetWindow()->GetHandle());
+        ofn.lpstrFile = szFile;
+        ofn.nMaxFile = sizeof(szFile);
+        if (GetCurrentDirectoryA(256, currentDir))
+            ofn.lpstrInitialDir = currentDir;
+
+        ofn.lpstrFilter = filter;
+        ofn.nFilterIndex = 1;
+        ofn.Flags = OFN_PATHMUSTEXIST | OFN_FILEMUSTEXIST | OFN_NOCHANGEDIR;
+
+        if (GetOpenFileNameA(&ofn) == TRUE)
+            return ofn.lpstrFile;
+
+        return std::string();
+    }
+
+    std::string Platform::SaveFileDialog(const char *filter)
+    {
+        OPENFILENAMEA ofn;
+        CHAR szFile[260] = {0};
+        CHAR currentDir[256] = {0};
+        ZeroMemory(&ofn, sizeof(OPENFILENAME));
+        ofn.lStructSize = sizeof(OPENFILENAME);
+        ofn.hwndOwner = glfwGetWin32Window((GLFWwindow *)Engine::GetWindow()->GetHandle());
+        ofn.lpstrFile = szFile;
+        ofn.nMaxFile = sizeof(szFile);
+        if (GetCurrentDirectoryA(256, currentDir))
+            ofn.lpstrInitialDir = currentDir;
+        ofn.lpstrFilter = filter;
+        ofn.nFilterIndex = 1;
+        ofn.Flags = OFN_PATHMUSTEXIST | OFN_OVERWRITEPROMPT | OFN_NOCHANGEDIR;
+
+        // Sets the default extension by extracting it from the filter
+        ofn.lpstrDefExt = strchr(filter, '\0') + 1;
+
+        if (GetSaveFileNameA(&ofn) == TRUE)
+            return ofn.lpstrFile;
+
+        return std::string();
+    }
+
 }
 
 #endif